Renaming the APIs to its original naming scheme. And creating files for each API and rename all temporarily created ViewModels to their original naming scheme. Fixing the problem with Cafeterias, which do not have any menus and the API is returning a 404 error. Now just a plain 'No Menus available' text shows up instead of an error message. Fixing a problem with the PersonSearch where the pIdentNr for the API was the wrong number, so the response was an error, now the API gets called with the obfuscated_id and the right results are retrieved.
