+++ Campus-iOS/GradesComponent/Service/GradesService.swift
@@ -10,7 +10,7 @@ import Alamofire
 
 struct GradesService: ServiceTokenProtocol {
     func fetch(token: String, forcedRefresh: Bool = false) async throws -> [Grade] {
-        let response: TUMOnlineAPI2.Response<Grade> = try await MainAPI.makeRequest(endpoint: TUMOnlineAPI2.personalGrades, token: token, forcedRefresh: forcedRefresh)
+        let response: TUMOnlineAPI.Response<Grade> = try await MainAPI.makeRequest(endpoint: TUMOnlineAPI.personalGrades, token: token, forcedRefresh: forcedRefresh)
         
         return response.row
     }
