+++ Campus-iOS/LectureComponent/Service/LecturesService.swift
@@ -15,9 +15,9 @@ protocol LecturesServiceProtocol {
 
 struct LecturesService: LecturesServiceProtocol {
     func fetch(token: String, forcedRefresh: Bool = false) async throws -> [Lecture] {
-        let response: TUMOnlineAPI2.Response<Lecture> =
+        let response: TUMOnlineAPI.Response<Lecture> =
         try await
-        MainAPI.makeRequest(endpoint: TUMOnlineAPI2.personalLectures, token: token, forcedRefresh: forcedRefresh)
+        MainAPI.makeRequest(endpoint: TUMOnlineAPI.personalLectures, token: token, forcedRefresh: forcedRefresh)
         
         return response.row
     }
