+++ Campus-iOS/LectureSearchComponent/Service/LectureSearchService.swift
@@ -9,7 +9,7 @@ import Foundation
 
 struct LectureSearchService {
     func fetch(for query: String, token: String, forcedRefresh: Bool) async throws -> [Lecture] {
-        let response : TUMOnlineAPI2.Response<Lecture> = try await MainAPI.makeRequest(endpoint: TUMOnlineAPI2.lectureSearch(search: query), token: token, forcedRefresh: forcedRefresh)
+        let response : TUMOnlineAPI.Response<Lecture> = try await MainAPI.makeRequest(endpoint: TUMOnlineAPI.lectureSearch(search: query), token: token, forcedRefresh: forcedRefresh)
         
         return response.row
     }
