+++ Campus-iOS/MapComponent/Service/StudyRoomsService.swift
@@ -9,13 +9,13 @@ import Foundation
 
 struct StudyRoomsService {
     func fetch(forcedRefresh: Bool) async throws -> StudyRoomApiRespose {
-        let response: StudyRoomApiRespose = try await MainAPI.makeRequest(endpoint: TUMDevAppAPI2.rooms, forcedRefresh: forcedRefresh)
+        let response: StudyRoomApiRespose = try await MainAPI.makeRequest(endpoint: TUMDevAppAPI.rooms, forcedRefresh: forcedRefresh)
         
         return response
     }
     
     func fetchMap(room: String, forcedRefresh: Bool) async throws -> [RoomImageMapping] {
-        let response: [RoomImageMapping] = try await MainAPI.makeRequest(endpoint: TUMCabeAPI2.roomMaps(room: room))
+        let response: [RoomImageMapping] = try await MainAPI.makeRequest(endpoint: TUMCabeAPI.roomMaps(room: room))
         
         return response
     }
