+++ Campus-iOS/MoviesComponent/Service/MovieService.swift
@@ -10,7 +10,7 @@ import Foundation
 struct MoviesService: ServiceProtocol {
     func fetch(forcedRefresh: Bool = false) async throws -> [Movie] {
         
-        let response: [Movie] = try await MainAPI.makeRequest(endpoint: TUMCabeAPI2.movie, forcedRefresh: forcedRefresh)
+        let response: [Movie] = try await MainAPI.makeRequest(endpoint: TUMCabeAPI.movie, forcedRefresh: forcedRefresh)
         
         return response
     }
