+++ Campus-iOS/PersonSearchComponent/Service/PersonSearchService.swift
@@ -9,7 +9,7 @@ import Foundation
 
 struct PersonSearchService {
     func fetch(for query: String, token: String, forcedRefresh: Bool) async throws -> [Person] {
-        let response : TUMOnlineAPI2.Response<Person> = try await MainAPI.makeRequest(endpoint: TUMOnlineAPI2.personSearch(search: query), token: token, forcedRefresh: forcedRefresh)
+        let response : TUMOnlineAPI.Response<Person> = try await MainAPI.makeRequest(endpoint: TUMOnlineAPI.personSearch(search: query), token: token, forcedRefresh: forcedRefresh)
         
         return response.row
     }
